# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for WOD Strategist application.",
        "title": "WOD Strategist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/{session_id}/activity-summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "Get observed workout activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Owned profile ID",
                        "name": "profile_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/activity.Summary"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "activity.Gap": {
            "type": "object",
            "properties": {
                "chunk_id": {
                    "type": "integer"
                },
                "clock": {
                    "type": "string"
                },
                "end_secs": {
                    "type": "number"
                },
                "reason": {
                    "type": "string"
                },
                "start_secs": {
                    "type": "number"
                }
            }
        },
        "activity.Interval": {
            "type": "object",
            "properties": {
                "end_secs": {
                    "type": "number"
                },
                "reason": {
                    "type": "string"
                },
                "start_secs": {
                    "type": "number"
                }
            }
        },
        "activity.Observation": {
            "type": "object",
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "end_secs": {
                    "type": "number"
                },
                "evidence": {
                    "type": "string"
                },
                "movement": {
                    "type": "string"
                },
                "start_secs": {
                    "type": "number"
                },
                "unit": {
                    "description": "reps or seconds",
                    "type": "string"
                }
            }
        },
        "activity.Observations": {
            "type": "object",
            "properties": {
                "activity_state": {
                    "description": "exercise, rest, unknown",
                    "type": "string"
                },
                "duration_secs": {
                    "type": "number"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/activity.Observation"
                    }
                },
                "target_state": {
                    "description": "identified, ambiguous, not_visible, unknown",
                    "type": "string"
                },
                "unassessed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/activity.Interval"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "activity.ReviewedChunk": {
            "type": "object",
            "properties": {
                "chunk_id": {
                    "type": "integer"
                },
                "observations": {
                    "$ref": "#/definitions/activity.Observations"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "activity.Summary": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "coverage_scope": {
                    "type": "string"
                },
                "media_generation": {
                    "type": "string"
                },
                "movements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/activity.Total"
                    }
                },
                "review_state": {
                    "type": "string"
                },
                "review_version": {
                    "type": "integer"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/activity.ReviewedChunk"
                    }
                },
                "source_version": {
                    "type": "string"
                },
                "unassessed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/activity.Gap"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "activity.Total": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "movement": {
                    "type": "string"
                },
                "seconds": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
    }
}